Add `setupDB` file to ease `self-hosting`

+++ src/events/dispatchers/disconnectDispatch.ts
@@ -22,14 +22,16 @@ export default async function disconnectDispatch(
 
     // Cleanup and exit process when the test client disconnects
     if (variables.testMode) {
-        // Delete all collections
-        await variables.prismaClient.account.deleteMany({});
-        await variables.prismaClient.post.deleteMany({});
-        await variables.prismaClient.community.deleteMany({});
-        await variables.prismaClient.communityMessage.deleteMany({});
-        await variables.prismaClient.token.deleteMany({});
-        await variables.prismaClient.log.deleteMany({});
-        await variables.prismaClient.report.deleteMany({});
+        if (!variables.setupMode) {
+            // Delete all collections
+            await variables.prismaClient.account.deleteMany({});
+            await variables.prismaClient.post.deleteMany({});
+            await variables.prismaClient.community.deleteMany({});
+            await variables.prismaClient.communityMessage.deleteMany({});
+            await variables.prismaClient.token.deleteMany({});
+            await variables.prismaClient.log.deleteMany({});
+            await variables.prismaClient.report.deleteMany({});
+        }
 
         process.exit();
     }
